Refactor CommunityDropDown into a proper dropdown
